refactor: ChannelSubscriptionControllerTest로부터 ChannelControllerTest 분리
