+++ backend/src/test/java/com/pickpick/controller/ChannelSubscriptionControllerTest.java
@@ -19,22 +19,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Sql({"/truncate.sql", "/channel.sql"})
 class ChannelSubscriptionControllerTest extends RestDocsTestSupport {
 
-    @DisplayName("구독 여부를 포함하여 채널을 조회한다.")
-    @Test
-    void findAll() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders
-                        .get("/api/channels")
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer 2"))
-                .andExpect(status().isOk())
-                .andDo(restDocs.document(
-                        responseFields(
-                                fieldWithPath("channels.[].id").type(JsonFieldType.NUMBER).description("채널 아이디"),
-                                fieldWithPath("channels.[].name").type(JsonFieldType.STRING).description("채널 이름"),
-                                fieldWithPath("channels.[].isSubscribed").type(JsonFieldType.BOOLEAN).description("채널 구독 여부")
-                        )
-                ));
-    }
-
     @DisplayName("멤버의 구독 중인 채널을 조회한다.")
     @Test
     void findAllSubscribedChannel() throws Exception {
